requirements.txt
numpy==2.3.0
qdrant-client==1.14.3
python-dotenv==1.1.0
